# openapi.yml
openapi: 3.0.3
info:
  title: Team Dev Frontend Server API
  description: |-
  version: 0.2.1
servers:
  - url: https://team-dev-frontend-server.herokuapp.com/
...
paths:
...
  /post/{id}/:
    patch:
      tags:
       - post
      summary: Edit a post 
      description: edit post by id
      operationId: editPost
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: 'The id of the post that is being edited.'
          required: true
          schema:
            type: string
      responses:
       '201':
         description: successful
       '401':
         description: Unuthorised
         content:
            application/json:
             schema:
               $ref: '#/components/schemas/Error'
       '400':
         description: bad request
         content:
            application/json:
             schema:
              $ref: '#/components/schemas/Error'
    delete:
      tags:
        - post
      summary: Delete a post by id
      description: delete post by id
      operationId: deletePost
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: 'get post by its id'
          required: true
          schema:
            type: string
      responses:
        '201':
         description: successful
         content:
           application/json:
            schema:
             $ref: '#/components/schemas/DeletePost'
        '401':
         description: authentication failed
         content:
            application/json:
             schema:
               $ref: '#/components/schemas/Error'
        '404':
         description: not found 
         content:
            application/json:
             schema:
              $ref: '#/components/schemas/Error'
        '500':
          description: error creating comment
          content:
            application/json:
             schema:
              $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
      bearerAuth:
        type: http
        scheme: bearer
        bearerFormat: JWT 
  schemas:
...
    DeletePost:
      type: object
      properties:
        status:
          type: string
        data:
          properties:
            deletedPost:
              properties:
                id:
                  type: integer
                  format: int64
                content:
                  type: string
                userId:
                  type: integer
                  format: int64
                createdAt:
                  type: string
                  format: string
                updatedAt:
                  type: string
                  format: string
            deletedComments:
              type: object
              properties:
                count:
                  type: integer
...
    Error:
      type: object
      properties:
        status:
          type: string
        data:
          properties:
            error:
              type: string
...